requirements.txt
Flask==3.0.3
flask_bcrypt==1.0.1
flask_caching==2.3.0
flask_limiter==3.8.0
flask_login==0.6.3
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
itsdangerous==2.2.0
Pillow==11.0.0
WTForms==3.2.1
